Rename `file_scope.rs` to `external_scope.rs`
