+++ crates/oak_ide/src/external_scope.rs
