+++ crates/oak_ide/src/lib.rs
@@ -1,8 +1,8 @@
-mod file_scope;
+mod external_scope;
 mod goto_definition;
 
 use biome_rowan::TextRange;
-pub use file_scope::ExternalScope;
+pub use external_scope::ExternalScope;
 pub use goto_definition::goto_definition;
 use url::Url;
